Move EmbeddedFieldAttributePropertyDrawer
