+++ crest/Assets/Crest/Crest/Scripts/Helpers/Attributes/EmbeddedFieldAttribute.cs
@@ -7,6 +7,10 @@ namespace Crest
     using Crest.EditorHelpers;
     using UnityEngine;
 
+#if UNITY_EDITOR
+    using UnityEditor;
+#endif
+
     public class EmbeddedFieldAttribute : PropertyAttribute
     {
         internal EmbeddedAssetEditor editor;
@@ -16,4 +20,19 @@ namespace Crest
             editor = new EmbeddedAssetEditor();
         }
     }
+
+#if UNITY_EDITOR
+    [CustomPropertyDrawer(typeof(EmbeddedFieldAttribute))]
+    public class EmbeddedFieldAttributePropertyDrawer : PropertyDrawer
+    {
+        public override void OnGUI(Rect position, SerializedProperty property, GUIContent label)
+        {
+            EmbeddedFieldAttribute embeddedAttribute = (EmbeddedFieldAttribute)attribute;
+            embeddedAttribute.editor.DrawEditorCombo(this, property, "asset");
+        }
+
+        // Removes space above embedded editor so the embedded editor replaces this drawer.
+        public override float GetPropertyHeight(SerializedProperty property, GUIContent label) => 0f;
+    }
+#endif
 }
