--- crest/Assets/Crest/Crest/Scripts/Helpers/Editor/EmbeddedAttributeEditor.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 491b77fa5c9714529986b4e2f8de91cb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
